Level 3 1 review

+++ src/components/SavedMovies/SavedMovies.js
@@ -9,50 +9,42 @@ import Preloader from "../Preloader/Preloader";
 function SavedMovies({openPopup}) {
     const [filmsArr, setFilmsArr] = useState([]);
     const [filmsInputSearch, setFilmsInputSearch] = useState('');
-    const [filmsSaved, setFilmsSaved] = useState([]);
 
     const [preloader, setPreloader] = useState(false);
     const [error, setError] = useState(false);
 
     const [filmsSwitch, setFilmsSwitch] = useState(true);
 
-    const [deleteClick, setDeleteClick,] = useState(false);
+    const [deleteClick, setDeleteClick] = useState(true);
+
 
     const filterShortFilm = (moviesToFilter) => moviesToFilter.filter((item) => item.duration <= 40);
 
-    async function onBookmarkClick(film) {
-        try {
-            await mainApi.deleteMovies(film._id);
-            const newSaved = await mainApi.getMovies();
-            setDeleteClick(!deleteClick)
-            setFilmsSaved(newSaved);
+    async function onBookmarkClick(film, isAdded) {
+        if (!isAdded) {
+            try {
+                await mainApi.deleteMovies(film._id);
+                debugger;
 
-        } catch (err) {
-            openPopup(`Ошибка удаления фильма!`, false);
+                let temp = filmsArr.filter(obj => obj._id != film._id);
+                setFilmsArr(temp);
+                localStorage.setItem('filmsSaved', JSON.stringify(temp))
+
+            } catch (err) {
+                openPopup(`Ошибка удаления фильма!`, false);
+            }
         }
     }
 
     useEffect(() => {
-        mainApi.getMovies()
-            .then((data) => {
-                setFilmsArr(data);
-            })
-            .catch((err) => {
-                openPopup({err}, false);
-            });
+        const localStorageFilmsSaved = localStorage.getItem('filmsSaved');
+
+        if (localStorageFilmsSaved) {
+            setFilmsArr(JSON.parse(localStorageFilmsSaved));
+        }
+
     }, [deleteClick]);
 
-
-    useEffect(() => {
-        mainApi.getMovies()
-            .then((data) => {
-                setFilmsArr(data);
-            })
-            .catch((err) => {
-                openPopup({err}, false);
-            });
-
-    }, [openPopup]);
 
     async function handleGetFilmsSwitch() {
         setFilmsSwitch(!filmsSwitch);
@@ -73,24 +65,22 @@ function SavedMovies({openPopup}) {
             let filmsFilter = filmsArray.filter(({nameRU}) => nameRU.toLowerCase().includes(filmsInputSearch.toLowerCase()));
 
 
-            if ( filmsFilter.length > 0) {
+            if (filmsFilter.length > 0) {
                 if (filmsSwitch) {
                     openPopup('Найдено фильмов: ' + filmsFilter.length, true)
+                    setFilmsArr(filmsFilter);
                 }
 
                 if (!filmsSwitch) {
                     openPopup('Найдено фильмов: ' + filterShortFilm(filmsFilter).length, true)
+                    setFilmsArr(filmsFilter);
                 }
 
-                setFilmsArr(filmsFilter);
 
             } else {
                 openPopup('Ничего не найдено', false)
                 setFilmsArr(filmsArr);
-                //alert(filmsArr);
             }
-
-            //setFilmsArr(filmsFilter);
 
         } catch (err) {
             openPopup(`Во время запроса произошла ошибка.       Попробуйте позже.`, false);
@@ -115,7 +105,7 @@ function SavedMovies({openPopup}) {
                 <MoviesCardList
                     films={filmsSwitch ? filmsArr : filterShortFilm(filmsArr)}
                     onBookmarkClick={onBookmarkClick}
-                    filmsSaved={filmsSaved}
+                    filmsSaved={filmsArr}
                 />
             )}
         </section>
@@ -123,3 +113,4 @@ function SavedMovies({openPopup}) {
 }
 
 export default SavedMovies;
+
